tests/wpt/web-platform-tests/offscreen-canvas/the-offscreen-canvas/initial.reset.path.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: initial.reset.path</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>initial.reset.path</h1>
<p class="desc">Resetting the canvas state resets the current path</p>


<script>
var t = async_test("Resetting the canvas state resets the current path");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

offscreenCanvas.width = 100;
ctx.rect(0, 0, 100, 50);
offscreenCanvas.width = 100;
ctx.fillStyle = '#f00';
ctx.fill();
_assertPixel(offscreenCanvas, 20,20, 0,0,0,0, "20,20", "0,0,0,0");

t.done();

});
</script>
